Refactor Branch wrapper class imports
