+++ src/commands/autocomplete.ts
@@ -1,5 +1,5 @@
 import yargs, { Arguments } from "yargs";
-import { Branch } from "../wrapper-classes";
+import Branch from "../wrapper-classes/branch";
 
 yargs.completion("completion", (current, argv) => {
   const branchArg = getBranchArg(current, argv);
@@ -30,7 +30,11 @@ function getBranchArg(current: string, argv: Arguments): string | null {
   const currentCommand = argv["_"].join(" ");
 
   // gt branch checkout --branch <branch_name>
-  if ((currentCommand.includes("branch checkout") || currentCommand.includes("b checkout")) && "branch" in argv) {
+  if (
+    (currentCommand.includes("branch checkout") ||
+      currentCommand.includes("b checkout")) &&
+    "branch" in argv
+  ) {
     // Because --branch is an option on the overall command, we need to check
     // the value of current to make sure that the branch argument is the
     // current argument being entered by the user.
@@ -46,12 +50,19 @@ function getBranchArg(current: string, argv: Arguments): string | null {
   // a substring of another command). Since we're dealing with a positional,
   // we also want to make sure that the current argument is the positional
   // (position 3).
-  if (argv["_"].includes("bco") && argv["_"].length <= 3 && typeof current === "string") {
+  if (
+    argv["_"].includes("bco") &&
+    argv["_"].length <= 3 &&
+    typeof current === "string"
+  ) {
     return current;
   }
 
   // gt upstack onto <branch_name>
-  if (currentCommand.includes("upstack onto") || currentCommand.includes("us onto")) {
+  if (
+    currentCommand.includes("upstack onto") ||
+    currentCommand.includes("us onto")
+  ) {
     // Again, since we're detailing with a positional, we want to make sure
     // that the current argument being entered is in the desired position,
     // i.e. position 4.
