+++ src/lib/utils/single_commit.ts
@@ -1,4 +1,5 @@
-import { Branch, Commit } from "../../wrapper-classes";
+import { Commit } from "../../wrapper-classes";
+import Branch from "../../wrapper-classes/branch";
 
 export function getSingleCommitOnBranch(branch: Branch): Commit | null {
   const commits = branch.getCommitSHAs();
