+++ src/wrapper-classes/git_stack_builder.ts
@@ -1,5 +1,6 @@
-import { AbstractStackBuilder, Branch } from ".";
+import { AbstractStackBuilder } from ".";
 import { MultiParentError, SiblingBranchError } from "../lib/errors";
+import Branch from "./branch";
 
 export default class GitStackBuilder extends AbstractStackBuilder {
   protected getBranchParent(branch: Branch): Branch | undefined {
