+++ src/wrapper-classes/index.ts
@@ -1,5 +1,4 @@
 import AbstractStackBuilder from "./abstract_stack_builder";
-import Branch from "./branch";
 import Commit from "./commit";
 import GitStackBuilder from "./git_stack_builder";
 import MetadataRef from "./metadata_ref";
@@ -12,7 +11,6 @@ export {
   AbstractStackBuilder,
   GitStackBuilder,
   MetaStackBuilder,
-  Branch,
   Commit,
   StackNode,
   MetadataRef,
