+++ test/fast/commands/stack/fix.test.ts
@@ -1,5 +1,5 @@
 import { expect } from "chai";
-import { Branch } from "../../../../src/wrapper-classes";
+import Branch from "../../../../src/wrapper-classes/branch";
 import { allScenes } from "../../../lib/scenes";
 import { configureTest, expectCommits } from "../../../lib/utils";
